controller: move errors to pkg errors
